Add files via upload
HEART DISEASE PREDICTION USING MACHINE LEARNING
The Goal of this Heart Disease Prediction is to understanding the heart disease study that produced
Cardiovascular heart disease prediction functions for assessing the risk of disorders of the heart
and the blood vessels. The dataset consist of the medical information based on the ECG report.
This Heart Disease Prediction will use Pandas, Numpy, Matplotlib, Scipy, Sklearn, and Seaborn
library for Prediction.

Dataset :-
This data set dates from 1988 and consists of four databases: Cleveland, Hungary, Switzerland,
and Long Beach V. It contains 76 attributes, including the predicted attribute, but all published
experiments refer to using a subset of 14 of them. The "target" field refers to the presence of heart
disease in the patient. It is integer valued 0 = no disease and 1 = disease.
Attribute Information:
1. age
2. sex
3. chest pain type (4 values)
4. resting blood pressure
5. serum cholestrol in mg/dl
6. fasting blood sugar > 120 mg/dl
7. resting electrocardiographic results (values 0,1,2)
8. maximum heart rate achieved
9. exercise induced angina
10. oldpeak = ST depression induced by exercise relative to rest
11. the slope of the peak exercise ST segment
12. number of major vessels (0-3) coloured by flouroscopy
13. thal: 0 = normal; 1 = fixed defect; 2 = reversable defect
 The names and social security numbers of the patients were recently removed
from the database, replaced with dummy values.

 https://www.kaggle.com/datasets/johnsmith88/heart-disease-dataset

 Implementation :-
Stage 1: Import Dependencies In this stage, we import the necessary libraries and
dependencies required for our heart disease prediction model. Some of the libraries we might need
include NumPy, Pandas, Matplotlib, Seaborn, and Scikit-learn.
Stage 2: Data collection and processing in this stage, we collect and process the heart
disease dataset. The dataset may contain various features such as age, sex, blood pressure,
cholesterol levels, and smoking habits. We clean the data by removing any duplicates, missing
values, or errors. We also normalize and scale the data to prepare it for machine learning algorithms.
Stage 3: Splitting the features and target in this stage, we split the data into input features
and target variables. The input features include age, sex, blood pressure, cholesterol levels, and
other medical conditions. The target variable is whether the individual has heart disease or not.
Stage 4: Algorithms stage in this stage, we train and test various machine learning
algorithms on our heart disease dataset. These might include logistic regression, MLP
Classification. We also perform hyperparameter tuning to optimize the performance of the
algorithms.
Stage 5: Accuracy score in this stage, we evaluate the performance of the machine learning
algorithms by calculating their accuracy score. This is done by comparing the predicted outcomes
with the actual outcomes in the test set. We can also calculate other metrics such as precision,
recall, and F1 score.
Stage 6: Building a predictive system In this stage, we build a predictive system using the
machine learning algorithm that performs the best in terms of accuracy score. This system takes
in input features, such as age, sex, blood pressure, cholesterol levels, and other medical conditions,
and predicts the likelihood of an individual developing heart disease. This predictive system can
be used by healthcare providers to identify high-risk patients and provide early intervention to
prevent or manage heart disease.